Three speed updates
Removed some of the logging.
Improved the ReadMe.
Introduced a third speed setting which will kick in after a number of slow schedule updates to further reduce bandwidth and demand on the car.
Because of the super slow updates, now any Alexa interaction will reset the updates to fast again