Interval logging fix
Added additional debug logging.
Fixed a bug where the 'interval' of the trigger would always be set to the slowUpdateTime even if the dormantUpdateTime was being set in the schedule.
Shouldn't affect functionality as there's no use for the interval value yet, however I did get some strange behaviour previously so will monitor