Nissan Connect failures output
If the battery status is checked and an invalid status is returned, the user is notified that there is a problem. This might happen when the car is out of range or similar, I couldn't get it to clear until I started the car and drove it around for a bit.